fix: improve files structure
